+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.23...master
 
+### Added
+- Add `COMMITED_MAIN` state
+
 ## [0.1.23] - 2021-02-09
 [0.1.23]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.22...release-0.1.23